move to doc::gen_html module
